requirements.txt
python-telegram-bot==21.6
requests==2.31.0
schedule==1.2.2
openai==1.51.0
